GPencil: Convert Blur size to float and avoid unedeed passes if size is 0

+++ source/blender/makesdna/DNA_shader_fx_types.h
@@ -81,7 +81,7 @@ typedef struct ShaderFxData_Runtime {
 
 typedef struct BlurShaderFxData {
   ShaderFxData shaderfx;
-  int radius[2];
+  float radius[2];
   /** Flags. */
   int flag;
   /** Number of samples. */
